Extract scheduling logic from FuturesUnordered
This commit extracts the scheduling logic from FuturesUnordered and
moves it to a private `scheduler` module. This module will be used by
the CurrentThread executor from the Tokio reform RFC.
